refactor: separate files for tools
